requirements.txt
 
streamlit
pymupdf
sentence-transformers
faiss-cpu
numpy
google-generativeai
requests
Pillow
pytesseract
python-dotenv